v0.22.5  Licensing command, improve help text, improve completion.

+++ centralcli/cache.py
@@ -137,6 +137,25 @@ class CentralObject:
             ]
         )
 
+    @property
+    def summary_text(self):
+        if self.is_site:
+            parts = [a for a in [self.name, self.city, self.state, self.zipcode] if a]
+        elif self.is_dev:
+            parts = [p for p in [self.name, self.serial, self.mac, self.ip, self.site] if p]
+            if self.site:
+                parts[-1] = f"Site:{parts[-1]}"
+        else:
+            return str(self)
+
+        return "|".join(
+            [
+                typer.style(p, fg="blue" if not idx % 2 == 0 else "cyan") for idx, p in enumerate(parts)
+            ]
+        )
+
+
+
 # TODO Not used yet refactor to make consistent Response object available for when using contents of cache to avoid API call
 class CacheResponses:
     def __init__(
@@ -203,7 +222,7 @@ class Cache:
     ) -> None:
         """Central-API-CLI Cache object
         """
-        self.rl: str = ""  # TODO temp might refactor cache updates to return resp
+        # self.rl: str = ""  # TODO temp might refactor cache updates to return resp
         self.updated: list = []  # TODO change from list of methods to something easier
         self.central = central
         self.responses = CacheResponses()
@@ -390,13 +409,13 @@ class Cache:
     def dev_switch_completion(
         self,
         incomplete: str,
-        args: List[str] = None,
+        args: List[str] = [],
     ):
         """Device completion for returning matches that are either switch or AP
 
         Args:
             incomplete (str): The last partial or full command before completion invoked.
-            args (List[str], optional): The previous arguments/commands on CLI. Defaults to None.
+            args (List[str], optional): The previous arguments/commands on CLI.
         """
         match = self.get_dev_identifier(incomplete, dev_type=["switch"], completion=True)
 
@@ -471,26 +490,79 @@ class Cache:
             for m in out:
                 yield m if isinstance(m, tuple) else (m, f"{ctx.info_name} ... {m}")
 
-    def dev_switch_ap_completion(
+    def dev_ap_completion(
         self,
         incomplete: str,
-        args: List[str] = None,
+        args: List[str] = [],
     ):
-        """Device completion for returning matches that are either switch or AP
+        """Completion for argument where only APs are valid.
 
         Args:
             incomplete (str): The last partial or full command before completion invoked.
             args (List[str], optional): The previous arguments/commands on CLI. Defaults to None.
         """
-        match = self.get_dev_identifier(incomplete, dev_type=["switch", "ap"], completion=True)
+        dev_types = ["ap"]
+        match = self.get_dev_identifier(incomplete, dev_type=dev_types, completion=True)
 
         out = []
         if match:
             for m in sorted(match, key=lambda i: i.name):
-                out += [tuple([m.name, m.help_text])]
+                if m.name not in args:
+                    out += [tuple([m.name, m.help_text])]
 
         for m in out:
-            yield m[0], m[1]
+            yield m
+
+    def dev_client_completion(
+        self,
+        incomplete: str,
+        args: List[str] = [],
+    ):
+        """Completion for client output.
+
+        Returns only devices that apply based on filter provided in command, defaults to clients
+        on both APs and switches (wires/wireless), but returns applicable devices if "wireless" or
+        "wired" filter is used.
+
+        Args:
+            incomplete (str): The last partial or full command before completion invoked.
+            args (List[str], optional): The previous arguments/commands on CLI.
+        """
+        gen = self.dev_switch_ap_completion
+
+        if args:
+            if args[-1].lower() == "wireless":
+                gen = self.dev_ap_completion
+            elif args[-1].lower() == "wired":
+                gen = self.dev_switch_completion
+            elif args[-1].lower() == "all":
+                return
+
+        for m in [dev for dev in gen(incomplete, args)]:
+            yield m
+
+    def dev_switch_ap_completion(
+        self,
+        incomplete: str,
+        args: List[str] = [],
+    ):
+        """Device completion for returning matches that are either switch or AP
+
+        Args:
+            incomplete (str): The last partial or full command before completion invoked.
+            args (List[str], optional): The previous arguments/commands on CLI.
+        """
+        match = self.get_dev_identifier(incomplete, dev_type=["switch", "ap"], completion=True)
+
+        # TODO fancy map to ensure dev.name, dev.mac, dev.serial, dev.ip are all not in args
+        out = []
+        if match:
+            for m in sorted(match, key=lambda i: i.name):
+                if m.name not in args:
+                    out += [tuple([m.name, m.help_text])]
+
+        for m in out:
+            yield m
 
     def dev_ap_gw_completion(
         self,
@@ -553,8 +625,7 @@ class Cache:
             args (List[str], optional): The previous arguments/commands on CLI. Defaults to None.
         """
         dev_types = ["ap", "gw"]
-        dev_match = self.get_dev_identifier(incomplete, dev_type=dev_types, completion=True)
-        match = [*self.get_group_identifier(incomplete, completion=True), *dev_match]
+        match = self.get_dev_identifier(incomplete, dev_type=dev_types, completion=True)
 
         out = []
         if match:
@@ -828,25 +899,23 @@ class Cache:
 
         return len(ret) == len(data)
 
-    # TODO have update methods return Response
+    # FIXME handle no devices in Central yet exception 837 --> cleaner.py 498
     async def update_dev_db(self):
         resp = await self.central.get_all_devicesv2()
         if resp.ok:
-            self.rl = str(resp.rl)
-            resp.output = utils.listify(resp.output)
-            resp.output = cleaner.get_devices(resp.output)
-            self.responses.dev = resp
-            # resp.output = [
-            #     {
-            #         k: v if k != "type" else get_cencli_devtype(v) for k, v in r.items()
-            #     } for r in resp.output
-            # ]
+            # self.rl = str(resp.rl)
+            if resp.output:
+                resp.output = utils.listify(resp.output)
+                resp.output = cleaner.get_devices(resp.output)
+
+                self.DevDB.truncate()
+                update_res = self.DevDB.insert_multiple(resp.output)
+                if False in update_res:
+                    log.error("Tiny DB returned an error during dev db update")
+
             # TODO change updated from  list of funcs to class with bool attributes or something
             self.updated.append(self.central.get_all_devicesv2)
-            self.DevDB.truncate()
-            update_res = self.DevDB.insert_multiple(resp.output)
-            if False in update_res:
-                log.error("Tiny DB returned an error during dev db update")
+            self.responses.dev = resp
         return resp
 
     async def update_site_db(self, data: Union[list, dict] = None, remove: bool = False) -> Union[List[int], Response]:
@@ -872,7 +941,7 @@ class Cache:
         else:
             resp = await self.central.get_all_sites()
             if resp.ok:
-                resp.output = utils.listify(resp.output)
+                resp.output = utils.listify(resp.output) if resp.output else []
                 resp.output = [{k.replace("site_", ""): v for k, v in d.items()} for d in resp.output]
                 self.responses.site = resp
                 # TODO time this to see which is more efficient
